DataLayoutMigration: Move flat XCSoarData files into typed subdirs
Run one-time migration after profile load: relocate top-level data files into typed folders, rewrite active profile paths, and leave a marker file. Keep the explicit Organize files action by sharing the same migration planner and profile-update logic.
